FIX: #if, #either & #switch belong to loader, not compiler FIX: remove #define directives at pre-processing stage FIX: #define & #include at compilation stage is an error
For conditional #define to work, the conditional directives (at the
moment: #if, #either, #switch) must be handled at the same stage at
which #define is handled. That is, they must be handled at
the earlier pre-processing (loading) stage and not at the later
compilation stage.

A simple example where the old behaviour causes problems:

    #switch OS [
        'Linux [#define libc-name "libc.so.6"]
        'Windows [#define libc-name "msvcrt.dll"]
    ]
    print libc-name

The pre-processor running on main.reds ignores the #switch and therefore
processes _both_ #define directives. `libc-name` is therefore _always_
defined to "libc.so.6" (as in the current pre-processor, redefinitions
are silently ignored), not conditionally based on the target OS, as
probably intended.

At compilation stage, the #switch is handled, but the #define in the
selected case is ignored (as #define is not handled by the compiler).

The first part of this commit fixes this problem, by moving the handling
of the conditional directives to the pre-processing stage.

The second part of this patch makes the pre-processor remove #define
directives from the code after they have been processed. The same was
already effectively done for #include directives, which are simply
replaced by the included file's contents.

With this in place, the compiler can be made more strict as to no longer
simply skip over #define and #include directives: any occurrence of a
pre-processing directive at compilation time is an error.