FEAT: use target-specific dynamic linker for ELF
This adds a `dynamic-linker` target option to configure which dynamic
linker ("interpreter") to use for ELF executables. This is intended to
make it easier to add other ELF-based platforms (e.g. FreeBSD) in the
future.

If no dynamic linker is configured, all interpreter-related information
is omitted from the binary. This leads to slightly more compact
executables for Syllable, which ignores the ELF interpreter completely
anyway (having the dynamic linker hard-wired in the kernel instead).