FIX: unless known otherwise, assume callbacks use cdecl
FIX: trivial cleanup

Some callbacks don't have their calling convention propagated at epilog
emission time, in which case we currently generate a "stdcall" epilog.
This patch switches this behaviour to generate a "cdecl" epilog per
default instead.

This is only a workaround. The real solution is to fix cconv
propagation.

This should be the final fix for issue #114.