FIX: deeper fix for cconv correct propagation to callbacks epilog.
The caller cconv propagation on callbacks function occurs on first callback invocation (using get-word! syntax). If done from global context, the invocation is compiled before the callback body. If done from a local context, the correct propagation depends on if the function containing the invocation is encountered, _before_ or _after_, the callback function in the source code. If the invocation context is compiled before, the propagation can happen, if after, the propagation happens after the callback compilation, so too late.

Based on this, the fix consists in deferring all callbacks body compilation after all non-callback functions to ensure that invocations are compiled before callbacks bodies. This doesn't cover the cases where callbacks would mutually call themselves (this construction could lead to infinite loops).

The case where a callback internal cconv has not been resolved (still equal to '??) when the body is about to be compiled, could be :

a) reported as a warning message
b) the callback body could be dropped
c) this case could be just ignored because if never called, the callback wrong epilog cannot be harmful.