FEAT: added new internal compiler event: on-root-level-entry
This event is notified to code back-end emitters when an expression or directive, at "root level", has been compiled. It allows to pin-point the location where a root expression ends in code buffer. It is currently only used by ARM literals pool manager, but will be useful also for a future source-level debugger.