FIX: issue #277 (Include system doesn't handle well files with same name in different directories)
This is a deep change in the Red/System loader, the include-dirs sub-system has been replaced by a stack-based approach. This should fix more issue than just this one. In order to make sure this fix doesn't break anything and, maybe, to push it further to fix more includes issues, it has been pushed in its own branch.