FIX: less aggressive context variable assignment catching. FIX: experimental extension of stack alignment to routines (disabled).
The context fix changes the following use case:
    var: 123
    A: context [
        foo: does [var: 0]
    ]
Before the fix, the compiler would create a A/var entry instead of binding the nested 'var to global context. Now, in such case, the user needs to explicitly declare 'var in the root scope of A to be part of that context.