FIX: crash on Linux 64bit caused by 32-bit integer overflow in bitset-tests.red.
Issue caused in test "basic-9" by the following expression:
    pick bs 2147483647

The internal index calculation leads to a wrong test for "virtual bit", trying then to read a not accessible memory address. The fact that this only happens on Linux 64bit and not any other platform suggests a race condition that is probably triggered by a different location for heap memory region on Linux 64bit, making it sensitive to this memory address calculation bug.