FEAT: preliminary support for dynamic calls support for paths. Fixes issue #926.
The compiler is now able to generate different code paths for path! values, one is passive evaluation, other one is trying to detect a function call at runtime (with possible refinements) and change the way subsequent expressions are stored on stack (accumulated instead of overwritten on stack frame base). Once the right number of arguments is available, the function call will be triggered and consume them.

This is a deep change, so regressions are possible.