FEAT: added first object's event: ON-WORD-SET* triggered just after an object's word path-setting operation.
The specification for implementing that event is:

    on-word-set*: func [word [word!] old [any-type!] new [any-type!]

It is supported both by the compiler and interpreter. The compiler will inline the call on object word's setting operations, when possible. Once defined, this callback is invoked for any object's word set using a path (external setting), so in-object direct words setting won't trigger it.

Usage example:

    a: object [
        b: 123
        show: does [self]
        on-word-set*: func [word old new][
            ?? word
            ?? old
            ?? new
            ;set word 789        ;-- changing the word's value is allowed
        ]
    ]
    a/b: 99