FEAT: aligned interpreter behavior to compiler's one wrt COPY on objects semantics.
COPY will reBIND object's functions to the new object (on contrary to Rebol3). Unfortunately, that is the only option we have for the Red compiler, due to implementation constraints. As methods calls take an object reference as argument internally in Red, the compiler needs to figure out accurately which object the method belongs to. Statically determining if an object has been COPYed or not is impossible in the general case, so the compiler needs to stick to the only reliable semantics it can handle, which is to use the newest object as context for the method calls.