FEAT: adds a new R/S compilation option: `debug-safe?`
If debug-safe? is set to yes (default: no), the runtime error reporting code will avoid dereferencing c-string and other memory pointers to avoid invalid memory accesses, resulting in an over-crashing. Use this option if the default call stack reporting crashes before the relevant part of the stack is printed. The option is not set by default to preserve printing c-string arguments in readable way.