FEAT: add on-wheel event for mouse wheel scrolling
The `event/count` property indicates the increment the wheel is rotated,
expressed in multiples of 1. A positive value indicates that the wheel was
rotated forward,away from the user; a negative value indicates that the
wheel was rotated backward, toward the user.

The `event/flags` property indicates whether various specail keys are
down. e.g. CTRL, SHIFT.