FEAT: generates hint values in PE files when possible.
Hint values in import sections are set to 0 by our linker. This patch will search for the imported libraries locally, extract the right hint value for each imported function and set it in import tables. This in theory would speed up executables loading, but in practice, the hint values correspond to a specific version of the lib, and those values tend to change a lot in newer versions. Also in practice, there is no noticeable startup time with right hint values. So, it seems to worth the extra code and linking time. Therefore, this patch is left on a separate branch for reference, in case such code is needed in the future.