FIX: `stat!` struct stack allocation could cause crashes on Linux/ARM platforms.
`stat!` struct is not fully specified for some targets like Linux. The struct was allocated on stack using `[stat! value]` which resulted in compiler inferring a size value too small for the full struct. Now the struct is allocated dynamically using 144 bytes, which is the largest found value for `stat!` struct.