FIX: crash in run-all-comp2 combined tests, compiled REPEAT redesigned.
This crash was caused by the corruption of the Red arguments stack series header (/offset and /tail overwritten). Code needed to be compiled in dev mode for the corruption to result in a crash. The corruption was caused by the combination of the `stack/arguments - 1` pattern (introduced by fix for #3361) in generated code for compiled REPEAT, a premature exit from the REPEAT loop using BREAK and the loop being at the root level in global space. In such case, the ending `stack/arguments - 1` is lower than `stack/bottom` causing the corruption of the stack series header.

Compiled REPEAT was redesigned entirely resulting in:
- shorter and much simpler code (both in the compiler and generated code).
- elimination of the need for global variables at R/S level to manager the counter.
- elimination of the hidden storage of counter word and value before the REPEAT arguments frame (thus eliminating the need to reach slots before `stack/arguments`).