FEAT: [R/S] more robust `loop` counter handling.
In R/S, the `loop` iterator was pushing the counter value on native stack, putting a burden on the body code to restore the stack exactly, in order for the counter to not be corrupted. This was creating tension with the exception handling system when `catch` keyword was used in the body.

This commit now injects one hidden local variable per `loop` keyword occurence, injecting the generated name (using a tag! converted to word! in order to avoid collision with user variables) in front of the `loop` keyword during the compilation. For loops in global space, a counter slot is allocated in the data segment (this option is not usable for loops in function because of possible recursions).