FEAT: adds `bounds` option to /options facet for restricted dragging.
Face/options/bounds, if set, can restrict the movements of the face when dragged inside a bounding box defined as:

    object [min: pair! max: pair!]

`min` refers to the top-left corner, `max` to the bottom-right corner.

Alternatively, a more dynamic bounding box can be returned by `on-drag-start` event handler, which can be changed on each new dragging session.