FIX: Crash during GC pass in unit tests.
The issue is related to an error! object generated during the tests, where the `arg3` field is set to `value1 word, which lies into the context of an op! that has been GCed. The word's `ctx` is still set to a node pointer, but that node pointer has been freed when the op! was collected, so that `ctx/value` is null, which crashes the GC during the marking phase.

The fix consists in copying such words on the stack, binding them to the global context, avoiding this issue.